+++ packages/dash-table/src/dash-table/components/Table/props.ts
@@ -318,6 +318,7 @@ export interface IProps {
     style_as_list_view?: boolean;
     page_action?: TableAction;
     page_current?: number;
+    page_count?: number;
     page_size: number;
 
     style_data?: Style;