+++ packages/dash-table/src/dash-table/components/Table/style.ts
@@ -1,6 +1,7 @@
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faEyeSlash, faTrashAlt } from '@fortawesome/free-regular-svg-icons';
 import { faEraser, faPencilAlt, faSort, faSortDown, faSortUp } from '@fortawesome/free-solid-svg-icons';
+import { faAngleLeft, faAngleRight, faAngleDoubleLeft, faAngleDoubleRight } from '@fortawesome/free-solid-svg-icons';
 
 library.add(
     faEraser,
@@ -9,5 +10,9 @@ library.add(
     faSort,
     faSortDown,
     faSortUp,
-    faTrashAlt
+    faTrashAlt,
+    faAngleLeft,
+    faAngleRight,
+    faAngleDoubleLeft,
+    faAngleDoubleRight
 );